Allow slashes in job name
Some Jenkins instances are configured to name jobs with a path-like syntax (e.g. 'dsyed/job/example/job/master'). By default, Flask will treat a job name like that as a different path on the server. Telling Flask to expect a 'path' allows access to the entire job name as a string.